organise /lib so that all group related files are in /lib/group
