+++ src/lib/group/createGroupJoinId.ts
@@ -1,4 +1,4 @@
-import prisma from "./prisma";
+import prisma from "../prisma";
 
 export default async function createGroupJoinId(joiningGroupId:number, isMaster:boolean) {
   const groupJoinId = await prisma.groupJoinId.create({
